requirements.txt
click==6.6
hvac==0.2.16
pyhcl==0.2.2
requests==2.12.3
colorama==0.3.7
config==0.3.9
